add lab_3

+++ students/k33421/Afonina_Natalia/Lab_3/lab_3/hotel/models.py
@@ -27,7 +27,7 @@ class Room(models.Model):
     )
     number = models.IntegerField()
     category = models.CharField(max_length=6, choices=ROOM_CATEGORIES)
-    cost = models.DecimalField(max_digits=10, decimal_places=2)
+    cost = models.IntegerField()
     phone = models.CharField(max_length=15)
     floor = models.ForeignKey(Floor, on_delete=models.CASCADE)
